Revert "32bit警告対策"
This reverts commit d71dc69500fb2a01bf68dc2e0eda848644fc36d3.

+++ sakura_core/view/CEditView.cpp
@@ -1864,7 +1864,7 @@ bool CEditView::GetSelectedData(
 		//>> 2002/04/18 Azumaiya
 
 		// メモリ確保に失敗したら抜ける
-		if( buffer->Capacity() < size_t(nBufSize) ){
+		if( buffer->Capacity() < nBufSize ){
 			return false;
 		}
 
@@ -1948,7 +1948,7 @@ bool CEditView::GetSelectedData(
 		//>> 2002/04/18 Azumaiya
 
 		// メモリ確保に失敗したら抜ける
-		if( buffer->Capacity() < size_t(nBufSize) ){
+		if( buffer->Capacity() < nBufSize ){
 			return false;
 		}
